millis() overflow logic reworked

+++ include/reset.h
@@ -12,6 +12,6 @@
 void do_reset(bool warmstart);
 void do_after_reset(void);
 void enter_deepsleep(const uint64_t wakeup_sec, const gpio_num_t wakeup_gpio);
-uint64_t uptime(void);
+unsigned long long uptime(void);
 
 #endif // _RESET_H
